Support checksum comparison of destination
Now the library supports the following checksumming capabilities:

- Compare checksum of an existing file at the destination.
- Compare checksum of the file that was downloaded.
- Compare checksum of the processed file at the destination.

This is important because the destination path may have a different
checksum if the destination was processed in any way, such as if the
file that was downloaded was a compressed file that was decompressed
to the destination.

A case where this scenario will happen is when downloading a
compressed distribution file from an apt repository, which requires
comparing the checksum of the uncompressed destination.