feat!: Rewrite crate with async/await support
BREAKING CHANGE: The project has been reset as a clean slate to be
developed around the new async/await syntax.

The new API is much easier to use in practice, and should result in
higher efficiency. Tokio is no longer a dependency, as the project
has moved to using a combination of async-std and surf.