chore: fix build by changing rust-toolchain and surf versions
I'm not sure what to put in `rust-toolchain` other than just `stable`.
Pinning a particular version is liable to break without a `Cargo.lock`
file, since a dependency is likely to begin requiring a new rustc
version without a semver bump.

There are several newer alphas of surf 2.0.0, but I left it at the one
that works for now. At least once 2.0.0 is actually released, this
should be updated.